% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{AddCounty}
\alias{AddCounty}
\title{Adds county to dataset}
\usage{
AddCounty(data)
}
\arguments{
\item{data}{A data.table}
}
\description{
Takes a location (either a municipality or county).
If the location is a county, nothing happens.
If the location is a municipality, it searches for the encompassing county
and adds this to the data.table \code{data} as a new variable \code{county}.
}
