% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{AddLocationName}
\alias{AddLocationName}
\title{Adds LocationName to dataset}
\usage{
AddLocationName(data)
}
\arguments{
\item{data}{A data.table}
}
\description{
Takes a location (either a municipality or county).
Finds the pretty name and adds this to the data.table \code{data}
 as a new variable \code{locationName}.
}
