% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{AnalyseRecentLine}
\alias{AnalyseRecentLine}
\title{Run a daily analysis with correctly specified data}
\usage{
AnalyseRecentLine(data, v)
}
\arguments{
\item{data}{A data.table. TODO: Validate the input.}

\item{v}{The version of sykdomspuls to use. Does not work.}
}
\description{
This function runs a daily analysis where the data
has been correctly formatted for the desired analysis.
}
