% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesSecondary.R
\name{AnalyseStats1}
\alias{AnalyseStats1}
\title{AnalyseStats1}
\usage{
AnalyseStats1(resYearLine = NULL, resYearLineMunicip = NULL)
}
\arguments{
\item{resYearLine}{a}

\item{resYearLineMunicip}{a}
}
\description{
AnalyseStats1
}
