% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_bayesian.R
\name{bayesian_load_stack_schema}
\alias{bayesian_load_stack_schema}
\title{Create analysis stack and datasets for a tag}
\usage{
bayesian_load_stack_schema(conf, data, stack_x)
}
\arguments{
\item{conf}{A row from \code{CONFIG$SYNDROMES}}

\item{data}{a dataset}

\item{stack_x}{a schema}
}
\description{
Given one row from \code{CONFIG$SYNDROMES} we need
to generate an analysis stack and all relevant datasets
that can be directly sent to \code{QuasipoissonTrainPredictData}
without additional formatting.
}
