% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{calculate_completeness}
\alias{calculate_completeness}
\title{Calculate completeness}
\usage{
calculate_completeness(location, x_year, x_granularity_time = "weekly",
  table = NULL)
}
\arguments{
\item{granularity_time}{}
}
\description{
Calculate completeness
}
