% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{calculate_confidence_interval}
\alias{calculate_confidence_interval}
\title{Calculate Confidence Interval}
\usage{
calculate_confidence_interval(data, last_weeks = NULL)
}
\arguments{
\item{granularity_time}{}
}
\description{
Calculate Confidence Interval
}
