% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{CONFIG}
\alias{CONFIG}
\title{Variables produced by the daily algorithm}
\format{A vector with one variable:
\describe{
  \item{date}{Date of the observation}
}
Configuration of analyses

This environment holds a number of important variables for
configuring the analyses that will be run.}
\usage{
CONFIG
}
\description{
A variable containing the variables produced by the daily algorithm.
The variabels are as follows:
}
\keyword{datasets}
