% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertConfigForAPI.R
\name{ConvertConfigForAPI}
\alias{ConvertConfigForAPI}
\title{Convert the new \code{CONFIG} for use in the old external API}
\usage{
ConvertConfigForAPI()
}
\description{
Convert the new \code{CONFIG} for use in the old external API
}
