% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateData.R
\name{DeleteLatestDoneFile}
\alias{DeleteLatestDoneFile}
\title{Delete the latest done file}
\usage{
DeleteLatestDoneFile(file = fhi::DashboardFolder("data_clean", paste0("done_",
  LatestRawID(), ".txt")))
}
\arguments{
\item{file}{Location of the latest done file (according to latest raw data file)}
}
\description{
Delete the latest done file
}
