% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{DeleteOldDatasets}
\alias{DeleteOldDatasets}
\title{Delete old datasets}
\usage{
DeleteOldDatasets(raw = list.files(fhi::DashboardFolder("data_raw"),
  "^partially_formatted_"),
  clean = list.files(fhi::DashboardFolder("data_clean"), "done_"))
}
\arguments{
\item{raw}{Folder containing raw data}

\item{clean}{Folder containing clean data}
}
\description{
Delete all old raw/clean datasets except the latest
}
