% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{displayDays}
\alias{displayDays}
\title{The last date for each isoweek}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 1567 rows and 2 columns.}
\usage{
displayDays
}
\description{
The last date for each isoweek
}
\keyword{datasets}
