% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailExternal}
\alias{EmailExternal}
\title{Sends an external email warning about alters}
\usage{
EmailExternal(results = readRDS(fhi::DashboardFolder("results",
  "outbreaks_alert_external.RDS")),
  alerts = readxl::read_excel(file.path("/etc", "gmailr",
  "emails_sykdomspuls_alert.xlsx")), isTest = TRUE, forceNoOutbreak = FALSE,
  forceYesOutbreak = FALSE)
}
\arguments{
\item{results}{Results file.}

\item{alerts}{Excel file containing the emails.}

\item{isTest}{Is this a test?}

\item{forceNoOutbreak}{For testing. Do you want to force a "No outbreak" email?}

\item{forceYesOutbreak}{For testing. Do you want to force a "Yes outbreak" email?}
}
\description{
Sends an external email warning about alters
}
