% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailInternal}
\alias{EmailInternal}
\title{Internal email for possible outbreaks}
\usage{
EmailInternal(resYearLine = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))))
}
\arguments{
\item{resYearLine}{Data from an resYearLine.RDS file}
}
\description{
Internal email for possible outbreaks
}
