% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNorMOMOInfluensa}
\alias{EmailNorMOMOInfluensa}
\title{Email notification of comparison between NorMOMO and Influensa}
\usage{
EmailNorMOMOInfluensa()
}
\description{
Email notification of comparison between NorMOMO and Influensa
}
