% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationOfFailedResults}
\alias{EmailNotificationOfFailedResults}
\title{Email notification of failed results}
\usage{
EmailNotificationOfFailedResults()
}
\description{
Email notification of failed results
}
