% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationOfNewResults}
\alias{EmailNotificationOfNewResults}
\title{Email notification of new results}
\usage{

  EmailNotificationOfNewResults(lastEmailedUtbruddFile = fhi::DashboardFolder("results",
  "lastEmailedUtbrudd.RDS"))
}
\arguments{
\item{lastEmailedUtbruddFile}{File containing the last week that emails were sent out}
}
\description{
Email notification of new results
}
