% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailTechnicalNewResults}
\alias{EmailTechnicalNewResults}
\title{Internal email notifying about new results}
\usage{
EmailTechnicalNewResults()
}
\description{
Internal email notifying about new results
}
