% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{FormatData}
\alias{FormatData}
\title{Format the raw data}
\usage{
FormatData(d, SYNDROME, population = GetPopulation(),
  hellidager = fread(system.file("extdata", "DatoerMedHelligdager.txt",
  package = "sykdomspuls"))[, c("Dato", "HelligdagIndikator"), with = FALSE],
  testIfHelligdagIndikatorFileIsOutdated = TRUE,
  removeMunicipsWithoutConsults = FALSE)
}
\arguments{
\item{d}{Raw data}

\item{SYNDROME}{Syndrome of interest}

\item{population}{Population file}

\item{hellidager}{Hellidager file}

\item{testIfHelligdagIndikatorFileIsOutdated}{Boolean. Test if the current date is older than the last hellidag recorded in the fiel?}

\item{removeMunicipsWithoutConsults}{Boolean. Remove municipalities that do not have any consultations?}
}
\description{
Format the raw data
}
