% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateData.R
\name{GenerateAnalysisResults}
\alias{GenerateAnalysisResults}
\title{Generate fake analysis results}
\usage{
GenerateAnalysisResults(granularity = "weekly", loc = "Norge",
  type = CONFIG$SYNDROMES_ALERT_EXTERNAL[1])
}
\arguments{
\item{granularity}{"daily"/"weekly"}

\item{loc}{Location code}

\item{type}{Syndrome}
}
\description{
Generate fake analysis results
}
