% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{GenerateFakeDataAnalysis}
\alias{GenerateFakeDataAnalysis}
\title{Generate fake analysis results}
\usage{
GenerateFakeDataAnalysis(syndrome = "influensa", xage = "Totalt",
  xmunicipEnd = "municip5054")
}
\arguments{
\item{syndrome}{Syndrome}

\item{xage}{Age}

\item{xmunicipEnd}{municipality}
}
\description{
Generate fake analysis results
}
