% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{GenerateFakeResultsFull}
\alias{GenerateFakeResultsFull}
\title{Generate fake analysis results}
\usage{
GenerateFakeResultsFull(granularity = "weekly", syndrome = "influensa",
  xage = "Totalt", xmunicipEnd = "municip5054")
}
\arguments{
\item{granularity}{daily/weekly}

\item{syndrome}{Syndrome}

\item{xage}{Age}

\item{xmunicipEnd}{municipality}
}
\description{
Generate fake analysis results
}
