% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateOutbreakListInternal.R
\name{GenerateOutbreakListInternal}
\alias{GenerateOutbreakListInternal}
\title{GenerateOutbreakListInternal}
\usage{
GenerateOutbreakListInternal(df = readRDS(fd::path("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fd::path("results", sprintf("\%s/resYearLineMunicip.RDS",
  LatestRawID()))), saveFiles = c(fd::path("results",
  sprintf("\%s/outbreaks.RDS", LatestRawID()))), useType = FALSE)
}
\arguments{
\item{df}{a}

\item{dk}{a}

\item{saveFiles}{a}

\item{useType}{a}
}
\description{
GenerateOutbreakListInternal
}
