% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{GenFakeDataAnalysis}
\alias{GenFakeDataAnalysis}
\title{Generate fake data for analysis}
\usage{
GenFakeDataAnalysis(syndrome = "influensa", xage = "Totalt",
  xmunicipEnd = "municip5054")
}
\arguments{
\item{syndrome}{Syndrome}

\item{xage}{Age}

\item{xmunicipEnd}{municipality}
}
\description{
Generate fake data for analysis
}
