% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{GenFakeDataClean}
\alias{GenFakeDataClean}
\title{Generate fake clean data}
\usage{
GenFakeDataClean(syndrome = "influensa", xmunicipEnd = "municip5054")
}
\arguments{
\item{syndrome}{Syndrome to validate}

\item{xmunicipEnd}{municipality}
}
\description{
Generate fake clean data
}
