% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{GenFakeDataRaw}
\alias{GenFakeDataRaw}
\title{Generate fake raw data}
\usage{
GenFakeDataRaw(xmunicipEnd = "municip5054")
}
\arguments{
\item{xmunicipEnd}{municipality}
}
\description{
Generate fake raw data
}
