% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorwayLocations.R
\name{GenNorwayMunicipMerging}
\alias{GenNorwayMunicipMerging}
\title{Creates the Norway Municip Merging (kommunesammenslaping) data.table}
\usage{
GenNorwayMunicipMerging()
}
\description{
Last updated 2017-07-29
}
