% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulation.R
\name{GenPopulation}
\alias{GenPopulation}
\title{Creates the population dataset
https://www.ssb.no/en/statbank/table/07459/tableViewLayout1/}
\usage{
GenPopulation()
}
\description{
Creates the population dataset
https://www.ssb.no/en/statbank/table/07459/tableViewLayout1/
}
