% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAlertsEmails.R
\name{GetAlertsEmails}
\alias{GetAlertsEmails}
\title{Get the correct alerts email excel sheet}
\usage{
GetAlertsEmails()
}
\description{
If this function is run on a production machine,
then the production email file will be returned.
Otherwise it will return the test file.
}
