% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{GetCountyFromMunicip}
\alias{GetCountyFromMunicip}
\title{Finds county from municipality}
\usage{
GetCountyFromMunicip(location, locationData = norwayLocations())
}
\arguments{
\item{location}{A location (either a municipality or county)}

\item{locationData}{Dataset containing a map between locatino code and pretty location name}
}
\description{
Takes a location (either a municipality or county).
If the location is a county, the county is returned.
If the location is a municipality, it searches for the encompassing county
and this is returned.
}
