% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses.R
\name{GetLocationName}
\alias{GetLocationName}
\title{Get location name from location code}
\usage{
GetLocationName(location, norwayLocations)
}
\arguments{
\item{location}{Location code}

\item{norwayLocations}{Dataset containing a map between locatino code and pretty location name}
}
\description{
When given a location code, the pretty location name is returned
}
