% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulation.R
\name{GetPopulation}
\alias{GetPopulation}
\title{GetPopulation
Mostly a function used by the package maintainer
to generate new population files as necessary}
\usage{
GetPopulation()
}
\description{
GetPopulation
Mostly a function used by the package maintainer
to generate new population files as necessary
}
