% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateData.R
\name{IdentifyDatasets}
\alias{IdentifyDatasets}
\title{Identify the latest raw/clean datasets}
\usage{
IdentifyDatasets(raw = list.files(fhi::DashboardFolder("data_raw"),
  "^partially_formatted_"),
  clean = list.files(fhi::DashboardFolder("data_clean"), "done_"))
}
\arguments{
\item{raw}{Folder containing raw data}

\item{clean}{Folder containing clean data}
}
\description{
Identify the latest raw/clean datasets
}
