% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateData.R
\name{LatestDatasets}
\alias{LatestDatasets}
\title{Latest datasets}
\usage{
LatestDatasets(clean = LatestRawID(), SYNDROME = "influensa")
}
\arguments{
\item{clean}{A date/ID}

\item{SYNDROME}{Syndrome that you are interested in}
}
\description{
Returns the names of the four datasets corresponding to a
date/syndrome.
}
