% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_analysis_list.R
\name{load_stack_schema}
\alias{load_stack_schema}
\title{Create analysis stack and datasets for a tag}
\usage{
load_stack_schema(conf, data, schema)
}
\arguments{
\item{conf}{A row from \code{CONFIG$SYNDROMES}}

\item{data}{a dataset}

\item{schema}{a schema}
}
\description{
Given one row from \code{CONFIG$SYNDROMES} we need
to generate an analysis stack and all relevant datasets
that can be directly sent to \code{QuasipoissonTrainPredictData}
without additional formatting.
}
