% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R, R/r6_mem.R
\docType{data}
\name{MeM}
\alias{MeM}
\title{Mem implementation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MeM

MeM
}
\description{
Mem implementation

Mem implementation
}
\keyword{datasets}
