% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mem.R
\docType{data}
\name{mem_schema}
\alias{mem_schema}
\title{mem_schema}
\format{An object of class \code{schema} (inherits from \code{R6}) of length 25.}
\usage{
mem_schema
}
\description{
DB schema for mem_results
}
\keyword{datasets}
