% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{models}
\alias{models}
\title{Model List}
\format{An object of class \code{list} of length 2.}
\usage{
models
}
\description{
This variable includes a list of all the model classes
}
\keyword{datasets}
