% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{norwayLocationsCreated}
\alias{norwayLocationsCreated}
\title{List of municipalities and counties}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 422 rows and 4 columns.}
\usage{
norwayLocationsCreated
}
\description{
List of municipalities and counties
}
\keyword{datasets}
