% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{norwayMunicipMergingCreated}
\alias{norwayMunicipMergingCreated}
\title{List of municipality merging over time}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5993 rows and 9 columns.}
\usage{
norwayMunicipMergingCreated
}
\description{
List of municipality merging over time
}
\keyword{datasets}
