% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{norwayPopulation}
\alias{norwayPopulation}
\title{Population}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 50554 rows and 5 columns.}
\usage{
norwayPopulation
}
\description{
Population
}
\keyword{datasets}
