% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{PB}
\alias{PB}
\title{Environment for progress bars}
\format{An object of class \code{environment} of length 2.}
\usage{
PB
}
\description{
Environment for progress bars
}
\keyword{datasets}
