% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB.R
\name{PBInitialize}
\alias{PBInitialize}
\title{Initialize the progress bar}
\usage{
PBInitialize(n, flush = fhi::DashboardIsDev())
}
\arguments{
\item{n}{a}

\item{flush}{a}
}
\description{
Initialize the progress bar
}
