% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{QuasiPoission}
\alias{QuasiPoission}
\title{Model classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QuasiPoission
}
\description{
Model classes
}
\keyword{datasets}
