% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsAggregate.R
\name{ResultsAggregateApply}
\alias{ResultsAggregateApply}
\title{ResultsAggregateApply}
\usage{
ResultsAggregateApply(stackFiles = ResultsAggregateStack())
}
\arguments{
\item{stackFiles}{The file stack that chooses which files to aggregate}
}
\description{
ResultsAggregateApply
}
