% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{schema_and_data}
\alias{schema_and_data}
\title{Create analysis stack and datasets for a tag inside a list}
\usage{
schema_and_data(conf)
}
\arguments{
\item{conf}{A row from \code{CONFIG$SYNDROMES}}
}
\description{
Given one row from \code{CONFIG$SYNDROMES} we need
to generate an analysis stack and all relevant datasets
that can be directly sent to \code{QuasipoissonTrainPredictData}
without additional formatting into a list for use in \code{pbmcapply}
}
