% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasipoisson_diagnostics.R
\name{update_diagnostics}
\alias{update_diagnostics}
\title{update_diagnostics}
\usage{
update_diagnostics(diagnostics, conf, x)
}
\arguments{
\item{diagnostics}{Diagnostics results}

\item{conf}{configuration object}

\item{x}{analysis run}
}
\description{
Updates diagnostics results with common data
}
