% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R
\name{ValidateResultsFull}
\alias{ValidateResultsFull}
\title{Validate the analysis results}
\usage{
ValidateResultsFull(d)
}
\arguments{
\item{d}{Results data.tabled}
}
\description{
Validate the analysis results
}
