% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{variablesAlgorithmDaily}
\alias{variablesAlgorithmDaily}
\title{Variables produced by the daily algorithm}
\format{A vector with one variable:
\describe{
  \item{date}{Date of the observation}
}}
\usage{
variablesAlgorithmDaily
}
\description{
A variable containing the variables produced by the daily algorithm.
The variabels are as follows:
}
\keyword{datasets}
