% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{VARS}
\alias{VARS}
\title{Global variables used for defining formats of data structures}
\format{An environment containing 3 variables:
\describe{
  \item{REQ_DATA_RAW_STRUCTURAL}{Required columns for raw data that are structural in nature (i.e. these form the skeleton).}
  \item{REQ_DATA_RAW_OTHER}{Required columns for raw data that are not structural in nature and are not listed in `CONFIG$SYNDROMES`.}
  \item{REQ_DATA_RAW_ALL}{The combination of `REQ_DATA_RAW_STRUCTURAL` and `REQ_DATA_RAW_OTHER`}
}}
\usage{
VARS
}
\description{
This environment holds a number of variables that are used for
defining the formats of data structures.
}
\keyword{datasets}
