% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesSecondary.R
\name{AnalyseStats1}
\alias{AnalyseStats1}
\title{AnalyseStats1}
\usage{
AnalyseStats1(resYearLine = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  resYearLineMunicip = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLineMunicip.RDS", LatestRawID()))))
}
\arguments{
\item{resYearLine}{a}

\item{resYearLineMunicip}{a}
}
\description{
AnalyseStats1
}
