% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_post_analysis.R
\name{clean_post_analysis}
\alias{clean_post_analysis}
\title{clean_post_analysis}
\usage{
clean_post_analysis(res, schema)
}
\arguments{
\item{res}{a}

\item{schema}{a}
}
\description{
clean_post_analysis
}
