% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_mem.R
\name{create_plots}
\alias{create_plots}
\title{create MEM season plots}
\usage{
create_plots(conf, mem_schema = NULL)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{mem_schema}{mem schema}
}
\description{
create MEM season plots
}
