% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateData.R
\name{DeleteOldDatasets}
\alias{DeleteOldDatasets}
\title{Delete old datasets}
\usage{
DeleteOldDatasets(raw = list.files(fd::path("data_raw"),
  "^partially_formatted_"), clean = list.files(fd::path("data_clean"),
  "done_"))
}
\arguments{
\item{raw}{Folder containing raw data}

\item{clean}{Folder containing clean data}
}
\description{
Delete all old raw/clean datasets except the latest
}
