% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailExternal}
\alias{EmailExternal}
\title{Sends an external email warning about alters}
\usage{
EmailExternal(results = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/outbreaks_alert_external.RDS", LatestRawID()))),
  alerts = GetAlertsEmails(), forceNoOutbreak = FALSE,
  forceYesOutbreak = FALSE)
}
\arguments{
\item{results}{Results file.}

\item{alerts}{Excel file containing the emails.}

\item{forceNoOutbreak}{For testing. Do you want to force a "No outbreak" email?}

\item{forceYesOutbreak}{For testing. Do you want to force a "Yes outbreak" email?}
}
\description{
Sends an external email warning about alters
}
