% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailNotificationOfNewData}
\alias{EmailNotificationOfNewData}
\title{Email notification of new data}
\usage{
EmailNotificationOfNewData(files)
}
\arguments{
\item{files}{The raw data file that is being analysed}
}
\description{
Email notification of new data
}
