% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListOutbreaks.R
\name{GenerateOutbreakListExternal}
\alias{GenerateOutbreakListExternal}
\title{Generates the external outbreak list}
\usage{

  GenerateOutbreakListExternal(df = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLineMunicip.RDS", LatestRawID()))),
  saveFiles = fhi::DashboardFolder("results",
  sprintf("\%s/outbreaks_alert_external.RDS", LatestRawID())),
  alerts = GetAlertsEmails())
}
\arguments{
\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{saveFiles}{Where are you saving the outbreak lists?}

\item{alerts}{Excel sheet containing registration information}
}
\description{
Generates the external outbreak list
}
