% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_ListOutbreaks.R
\name{GenerateOutbreakListInternal}
\alias{GenerateOutbreakListInternal}
\title{Generates the internal outbreak list}
\usage{
GenerateOutbreakListInternal(df = readRDS(fd::path("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fd::path("results", sprintf("\%s/resYearLineMunicip.RDS",
  LatestRawID()))), saveFiles = c(fd::path("results",
  sprintf("\%s/outbreaks.RDS", LatestRawID()))), useType = FALSE)
}
\arguments{
\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{saveFiles}{Where are you saving the outbreak lists?}

\item{useType}{Use type or tag as variable name}
}
\description{
Generates the internal outbreak list
}
