% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListOutbreaks.R
\name{GenerateOutbreakListInternal}
\alias{GenerateOutbreakListInternal}
\title{Generates the internal outbreak list}
\usage{

  GenerateOutbreakListInternal(df = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLine.RDS", LatestRawID()))),
  dk = readRDS(fhi::DashboardFolder("results",
  sprintf("\%s/resYearLineMunicip.RDS", LatestRawID()))),
  saveFiles = c(fhi::DashboardFolder("results",
  sprintf("\%s/outbreaks.RDS", LatestRawID()))), useType = FALSE)
}
\arguments{
\item{df}{Data inside resYearLine.RDS}

\item{dk}{Data inside resYearLineMunicip.RDS}

\item{saveFiles}{Where are you saving the outbreak lists?}

\item{useType}{Use type or tag as variable name}
}
\description{
Generates the internal outbreak list
}
