% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateData.R
\name{IdentifyAllDatasets}
\alias{IdentifyAllDatasets}
\title{Identify all raw/clean datasets that are available}
\usage{
IdentifyAllDatasets(raw = list.files(fd::path("data_raw"),
  "^partially_formatted_"), clean = list.files(fd::path("data_clean"),
  "done_"))
}
\arguments{
\item{raw}{Folder containing raw data}

\item{clean}{Folder containing clean data}
}
\description{
Identify all raw/clean datasets that are available
}
