% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_quasipoisson.R
\docType{data}
\name{quasipoission}
\alias{quasipoission}
\title{quasipoission}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
quasipoission
}
\description{
quasipoission
}
\keyword{datasets}
