% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_mem.R
\name{run_all_mem}
\alias{run_all_mem}
\title{run_all_mem}
\usage{
run_all_mem(conf, mem_schema)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{mem_schema}{a}
}
\description{
Runs mem analysis for all available years for the whole
country and for each county
}
