% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{StackAndEfficientDataForAnalysis}
\alias{StackAndEfficientDataForAnalysis}
\title{StackAndEfficientDataForAnalysis}
\usage{
StackAndEfficientDataForAnalysis(conf)
}
\arguments{
\item{conf}{A row from \code{CONFIG$SYNDROMES}}
}
\description{
StackAndEfficientDataForAnalysis
}
