% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackIterator.R
\name{StackIterator}
\alias{StackIterator}
\title{Stack Iterator}
\usage{
StackIterator(stackStrata, stack, data, progressFunction)
}
\arguments{
\item{stackStrata}{a}

\item{stack}{a}

\item{data}{a}

\item{progressFunction}{a}
}
\description{
Stack Iterator
}
