% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCreationAndValidation.R, R/ValidateData.R
\name{ValidateAnalysisResults}
\alias{ValidateAnalysisResults}
\alias{ValidateAnalysisResults}
\title{Validate analysis results}
\usage{
ValidateAnalysisResults(d, granularity = "weekly")

ValidateAnalysisResults(d, granularity = "weekly")
}
\arguments{
\item{d}{Results data.table}

\item{granularity}{"daily"/"weekly"}
}
\description{
Validate analysis results

Validate the analysis results
}
