% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\docType{data}
\name{VARS}
\alias{VARS}
\title{Global variables used for defining formats of data structures}
\format{An environment containing 5 variables:
\describe{
  \item{REQ_DATA_RAW}{Required columns for raw data.}
  \item{REQ_DATA_CLEAN}{Required columns for clean data.}
  \item{REQ_DATA_ANALYSIS}{Required columns for data analysis.}
  \item{REQ_RESULTS_BASIC}{Required columns for the results (basic, right after the analysis is run)}
  \item{REQ_RESULTS_FULL}{Required columns for the full results (i.e. those that will be saved and used further)}
}}
\usage{
VARS
}
\description{
This environment holds a number of variables that are used for
defining the formats of data structures.
}
\keyword{datasets}
